webpack modifications
It turned out that having a node_modules in my project director from a previous
npm install that was run in outside docker was being used by the webpack docker
container since . was mounted. The problem was when I started clean this
node_modules directory was not available, and then webpack docker build
installed the node_modules, but then when the . was mounted again it masked the
node_modules that were installed.

If that sounds confusing it's because it was :) I guess the lesson here is to be
careful with volumes, and the potential for masking things that were already
installed in the docker container.

This commit sets up two explicit volumes for the webpack docker container:

- docnow: which is where webpack BundleTracker writes webpack-stats.json
- assets: the directory of source code that webpack needs to watch for changes