Initial login
The logic around the first login had been modified to return twice which
was causing HTTP reponses to be sent twice. I've relaxed the eslint rule
that was preventing the final return from being wrapper in an else
statement. If there's a cleaner way to do this logic while supporting the
eslint rule I'm all for it. But the rule does seem a bit draconian :)

fixes #14