Showing instance places by default & user can add places
This required a few changes in both the client and server: the API endopoints for places have been changed to "places" for the whole instance and "userPlaces" for users.
Database operations also needed adjustments, but simpler: instance places are prefixed with 'instance'.

NB There are quite a few CSS issues that I haven't really tackled until we have more from Alexandra.

Fixes #28