ElasticSearch + Docker
Added configuration for ElasticSearch to docker-compose. The elasticsearch
host to use can be configured in the environment with ES_HOST. The
docker-compose.yml does this using the hostname set up by Docker.

I also noticed that we weren't specifying the version of node past the major
version, which seemed to cause some problems loading 'multiparty'. Being
specific about 8.4.0 seemed to resolve this issue.