importFromSearch
createSearch now will just create a search object in elasticsearch and
return it. getSearch will get this object. importFromSearch will go
out and do a search on twitter and import the results into elasticsearch
There are also the shell of some tests for getting particular stuff back
from the index that haven't been filled in yet.