Wait for search
This commit introduces an inintial wait before attempting to fetch stats
for the search. This is to account for situations where a user has
returned to the trends screen and wants to explore a new hashtag. Without
this wait they will temporarily see the old results before they are refreshed.