remove p font-family
There is no need to repeat the font-family again on the intro paragraph since it
is Roboto throughout (defined at the App level). This declaration was
syntactically wrong which led to a different between demo.docnow.io and dev
environments because webpack was trying to correct the font-family when
running under NOD_ENV=production.