URLs Card
This commit includes a bunch of changes for the URLs card for #40. It also
includes a fix for errors related to geo_shape parsing errors due to the
indexes not being mapped correctly. This exposed an off by one error in the
twitter results that I would like to fix before merging.