Explore improvements
This commit includes improvements to lower the traffic to the server for
updated stats when the search is not being actively updated. It is implemented
by storing an active property to the search elasticsearch document. This
is propagated back out to the client. When a search is active the update button
changes to "updating" and will not respond to clicks.

Also included are small style changes to the search summary to highlight the
counts and dates, and the use of the moment javascript library to display
human readable local times instead of UTC.