bar chart font size
This is a small change to make sure the font size is set correctly
on the bar chart. Also the linear scale for the barchart was adjusted
to use scale that is the squareroot of the value. This has the effect of
making smaller values larger. The hope is that this will allow people to
see finer gradations in the data for hashtags with smaller values.
If it's more confusing that linear scale it will be easy to switch back.