Url Fetching
This commit includes a lot of changes that implement URL fetching using
Redis as both a work queue, and also as a database for the retrieved webpage
metadata.

When tweets are retrieved and sent off for storage in ElasticSearch the
extracted urls are pushed onto a list in Redis. The url fetching is performed
by a separate worker process which you can start with:

    npm run url-fetcher

The worker will only pop one url off the queue at a time, so it may be good to
start more than one in a production environment.