getTweetsWithUrl
Keep track of the tweet ids that mention a particular URL in Redis, and
provide a method on db to fetch tweets that mention a given URL in a
particular search. This is working towards adding the ability to view
tweets that mention a particular URL in the user interface.