Tweets Modal
This commit adds a modal for viewing tweets related to a particular URL.
Hopefully the modal can be used in other places where displaying a list
of tweets is needed. It uses Twitter's embed widget to display. When
displaying retweets the embed is itself put inside a container that
gives attribution to the user who is retweeting.