moved server
server.js was moved into src/commands so it can be built in dist for
production. There are two npm commands for running it in development and
production:

- npm run start:dev
- npm run start

start:dev will watch for changes to the server and client code, and restart the server and client as needed using hot-module-replacement.