select/deselect webpages
This commit adds the ability to select and deselect webpages in a search.
A webpage can be in three different states: non-yet-evaluated, selected or
deselected. A website that has been selected is of interest for an
archive, one that has been deselected is not of interest, and
not-yet-evaluated means a decision has not been made.

The selection/deselection decisions are stored in two separate sets in Redis.