Wayback Indicator
This commit adds a small Internet Archive logo to each webpage card
which is green if the webpage is archived in the Wayback Machine.
If there is a match the tooltip should display when the page was
last archived. The Wayback React component uses a new API call
/api/v1/wayback?url={url} which in turn uses the availability API call
at https://archive.org/wayback/available. Ideally the client side app
could call the Wayback API directly but it doesn't support CORS.

Remaining work includes adding SavePageNow functionality if the item is
not in the Wayback. Also, docnow should store the results of the lookups
in Redis so that page refreshes don't cause it to look them all up
again.