exclude retweets in tweets view
When exploring it is useful to see only original tweets so as to not
be overwhelmed with retweets of the same tweet. It also would be useful
to order the tweets by how much they have been retweeted. This commit
adds a ?includeRetweets parameter to the /search/{search-id}/tweets API
call to optionally include retweets.

This commit also includes a new command 'update-indexes' to refresh the
elasticsearch mappings without dropping and adding them again.

fixes #71