.env for dev/test/prod
This commit includes a src/env which will load the appropriate .env if
available for the various environments: test, development and
production.

I added teardown and migration to db.clear() which will now allow the
tests to run again.

This commit also includes some changes to get User working in
postgresql. I wanted the tests to work so I could get this working. More
work remains though to get the many to many joins working.