db.getTwitterUsers()
This commit updates the db.getTwitterUsers() function to use the
database instead of elasticsearch. It was soooooo satisfying to remove
the cryptic elasticsearch query. But this may need to be adjusted over
time since Twitter users are not yet being modeled during the ingest
apart from their JSON being stashed into the Tweet.json column along
with the rest of the tweet info.