Improved initial key setup
This commit adds error catching during the initial key handshake which
sends the administrator back to the settings page with a (hopefully) helpful
error message that helps fix the problem. These messages may need to be
adjusted over time, but are certainly better than the current 500
message which can be very unfriendly for people who are just setting up
the application.

In addition the activation message on the profile page was simplified a
bit. Users are only redirected to their profile on login when they are
not active.