Quota handling
This commit fixes a bug in the quota handling that wasn't evaluating the
total number of tweets by user properly. This was present since the
switch to Postgres. The result of the query was not being used properly
and the update to the search to flag it as inactive wasn't being
committed properly.

The commit also includes some additional indexes to the tweet table
which will hopefully speed up some queries involving the screenName and
the searchId. These were identified by using pghero
https://github.com/ankane/pghero which is a very simple web app for
viewing slow database queries. The docker-compose includes pghero
running on port 3001. I think this is ok since the docker configuration
is really meant for development environments since we have
docnow-ansible.