react-modal
There are three modals in the app, two of which use react-modal and one
which uses material-ui's modal. Ideally I think material-ui's would be
used in all three, but for the moment to keep things looking the same
with CloseModal and modal titles the Trash component has been converted
to use react-modal.