New StreamLoader
The v2 Twitter stream API delivers all results as part of a single
stream. Each tweet that comes down the pipe has a tag, which identifies
the searches it belongs to with a tag. This simplifies the StreamLoader
quite a bit since it just needs to listen to the single stream, and then
the stream is controlled by adding and removing stream rules using the
Twitter API. The Database.startStream() and Datbase.stopStream() methods
are used to start and stop the stream for a search. The StreamLoader
class is started up and it basically connects and waits for tweets to
process.

Note: this new functionality means the DocNow app uses the v2 API for
both searching and streaming Twitter data. More work will need to be
done to allow the DocNow app to use the /tweets/search/all endpoint for
historical tweets. That's next!