Use edsu/twitter-v2
As documented over in
https://github.com/HunterLarco/twitter-v2/issues/90 there is a bug in
twitter-v2 that is preventing searches with spaces in them from being
signed properly, which results in a 401 HTTP error. I've sent a PR with
a fix, athough it feels like a bug in the underlying oauth-1.0a library.
Until this is resolved or we use a different library to talk to Twitter
v2 we will have to use edsu/twitter-v2. If this persists for a while we
should probably consider moving it to docnow/twitter-v2