Performance improvements
db.getSearchSummary() was removed and db.getSearch() now takes an
optional includeSummary parameter which defaults to false. When set to
true it will call db.getSearchStats() for the search and add the results
to the search object that is returned.

In addition db.getSearchStats() will now cache results in Redis for 60
seconds to avoid repeated lookups to the database when there are recent
results.