resetTwitterSearch
Make sure to call resetTwitterSearch when leaving the Insights and
Collection views so that state.search is emptied out. If the search is
left in there then when the user visits /explore/ they will see the
previous search which won't make much sense since /explore/ is for
generating a new search.