Over quota messages
Attempts to activate a search when the user is over quota will display
an error message. A new piece of state `message` can be used to display
a message on any page. It is set using the message.setMessage() action.
The over quota error is triggered when a a PUT is made to the API to
update the collection. The error response that comes back is used to
call setMessage().