New tweet index
Add an index on table tweet for (search_id, screen_name) which should
speed up lookups to see if a user is in a large collection. This index
can also be used for looking up tweets by search_id alone which means
that the tweet (search_id) index can be removed.