search-loader changes
This commit includes lots of changes to get search-loader working properly. It also includes some changes to the UserList component to only display stats for searches that have been saved, not temporary exploratory searches. Also some minor changes to how dates and numbers are displayed in a few places which I noticed while using the app.