Academic
When settings keys on the /settings/ page are updated the docnow server side app will do a historical search behind the scenes to see if they have Academic Research Product Track access turned on. The settings table will have the academic property set accordingly.

The academic settings property is then used to determine how far back the user can search when activating a search. The default for non-academic apps is one week. Academic apps can search back to the beginning of Twitter in 2006.