Simplify running tests
The test suite runs automatically via a GitHub Action. It has been possible to run the test suite manually in a development environment but the instructions previously required you to connect to the Postgres Docker container, manually create a test database, and then connect to the web app, edit a file and run the tests.

Now instead you simply copy a file:

```
cp .env.test-sample .env.test
```

Replace the `CHANGE_ME` values in .env.test to the respective Twitter API credentials.

And then run the tests.

```
npm run test
```

The tests are run using `docker-compose run` and the test database is
created automatically. The instructions in the README.md have been
updated appropriately.