add from/to to search activation panel
This commit allows the user to specify the start and end times for the
search. The default start is the current time, and the default end is
one week from then. Both start and end are required.

Because start and end need to be adapted for search and stream, the
SearchJob has been adapted to store the appropriate values on
tweetsStart and tweetsEnd.

For example if the current date is Feb 1,
2022 and the user selects to start on Jan 15, 2022 and end on Feb 14,
2002 the search needs to go from Jan 15, 2022 to Feb 1, 2022 (not Feb
14, 2022).

The QuotaChecker was adapted to also stop any stream jobs that are
active past their stop time.

The start and end times are date times now instead of displaying as
dates.

fixes #174