SearchTerm tooltips
On the trending screen if you hover over a term it will display a
tooltip saying "Explore tweets with the hashtag '#example'" and related
for the other search types. I also lowered the css border-radius made
the terms less pill-like and a bit boxier.

After experimenting with some examples of putting the colored search
term examles into the text below the search box I decided not to
implement it until we've had time to design what we want that to look
like.

Closes #178