Search loader test needs to set endDate
The search-loader test needed to set the endDate on the query when
creating the search. This was causing searchStart never to work since it
was not set. The user interface does this automatically now, but the
test was never updated to reflect that new requirement. In addition
db.startSearch() was updated to return the newly created SearchJob or
null if there was a problem.

Fixes #188