Archive race condition
This fixes a race condition in the writing of JavaScript and JSON
payloads during archive creation. It was causing the writing of the
archive to hang occasionally, which manifested in the user interface and
the unit tests. In addition the package.json package-lock.json were
updated with an `npm update`.

Fixes #190