Video collection
When we were using the Twitter v1.1 API we were able to get MP4 URLs directly
from the API JSON to allow playing them in the Video cards in Explore
and Insights. However the MP4 URLs are not made available (yet) in the
new V2 API. https://twittercommunity.com/t/how-to-get-video-from-media-key/152449

To still allow playback in the DocNow app itself the tweet loader will
place tweets that have videos onto a Redis queue, and the url-fetcher
worker will hydrate the tweets (as many as 100 at a time) and insert the
MP4 and thumbnail URLs in the database for the tweet.

This is really a stopgap measure, because at some point Twitter are
going to turn off the v1.1 API. At that point we will need to look for
other sources for places, trends, and videos.

Fixes #180