ToS page and editor
This adds a new page for Terms of Service which displays the default
Terms of Service specified in userData/terms-of-service.md. When the
super-user views the terms they edit the underlying markdown and save
it, at which point it is persisted to the settings database table and
used from there on out.

This commit also includes a bugfix for `boolean` -> `bool` in PropTypes
across React components.

fixes #48