Adaptable search endpoint
When the app is not configured with academic research prodcut track keys
it will not try to search the '/all' endpoint even when asked to. Fixes

This commit also fixes a bug in server.twitter when performing a single
search (no paging). The problem was that the search callback was getting
called twice, the second time with no next token, when search was told
to just do one search (once: true). This was a problem for the
SearchLoader which was receiving that second callback with no next token
and flagging the search as complete, even though there were more
results to fetch.